Add Tuff rock type, misc cleanup
As a sedimentary rock that does not currently generate in layers, only in tuff rings. Ores are typical sedimentary ores + sulfur. Sulfur should be tuned down substantially if tuff is made to generate in normal rock layers.

If we are to have it as a rock layer, I'd say it could be a rare top layer in volcanic areas, maybe it being on the surface always signifies IgEx mid layer, IgIn bottom layer?

My other planned use case for tuff in worldgen would be having layers of it in stratovolcanoes, assuming I can get those working.