Changed fire start behavior, added flint and pyrite
Flint and Pyrite, a mid-game firestarter w/ 128 durability that lights like flint and steel but with only a 1/3 chance (configurable) per click to start a strong fire. Lights torches 100% of the time. Uses durability regardless of whether a fire is started, so ~40 uses are expected

Flint and Steel (and Flint and Pyrite and fire charges and flaming arrows) can be used to light fire pits from items on ground. All except Flint and Pyrite have a 100% chance to light regardless of kindling

Made flint and steel and fire charges play their respective sounds when used